add gui cli
