+++ simple_generation/simple_generation.py
@@ -20,9 +20,11 @@ from transformers import (
     DataCollatorWithPadding,
     GenerationConfig,
 )
+import gradio as gr
+from transformers import TextIteratorStreamer
+from threading import Thread
 
 from .utils import DistributedEvalSampler
-from .gui import build_gui
 
 logger = get_logger(__name__)
 
@@ -441,7 +443,113 @@ class SimpleGenerator:
 
     def gui(self, type: str = "chat", **generation_kwargs):
         """Start a GUI for the model."""
-        interface = build_gui(type, self, **generation_kwargs)
+
+        if type == "chat":
+
+            def _chat(message, history):
+                messages = list()
+                for user_prompt, model_response in history:
+                    messages.append({"role": "user", "content": user_prompt})
+                    messages.append({"role": "assistant", "content": model_response})
+                messages.append({"role": "user", "content": message})
+
+                tokenized_chat = self.tokenizer.apply_chat_template(
+                    messages,
+                    tokenize=True,
+                    add_generation_prompt=True,
+                    return_tensors="pt",
+                ).to(self.device)
+
+                streamer = TextIteratorStreamer(
+                    self.tokenizer,
+                    timeout=10.0,
+                    skip_prompt=True,
+                    skip_special_tokens=True,
+                )
+                current_generation_args = self._prepare_generation_args(
+                    **self.generation_kwargs
+                )
+
+                gen_args = dict(
+                    inputs=tokenized_chat,
+                    streamer=streamer,
+                    **current_generation_args,
+                )
+
+                t = Thread(target=self.model.generate, kwargs=gen_args)
+                t.start()
+                partial_message = ""
+                for new_token in streamer:
+                    if new_token != "<":
+                        partial_message += new_token
+                        yield partial_message
+
+            interface = gr.ChatInterface(
+                _chat,
+                # chatbot=gr.Chatbot(height=300),
+                title=f"Chat with {self.model_name_or_path.split('/')[-1]}",
+                description="Generation arguments: " + str(self.generation_kwargs),
+                # fill_vertical_space=True, # this needs an upcoming gradio release
+            )
+
+        elif type == "translation":
+
+            with gr.Blocks() as interface:
+
+                def _run_translation(src_lang, src_text, tgt_lang):
+                    texts = [s.strip() for s in src_text.split("\n")]
+
+                    return self(
+                        texts,
+                        skip_prompt=False,
+                        forced_bos_token_id=self.tokenizer.lang_code_to_id["ita_Latn"],
+                        **generation_kwargs,
+                    )
+
+                with gr.Row():
+                    gr.Markdown("### Simple Generation Translation Interface")
+
+                with gr.Row():
+                    with gr.Column():
+                        src_lang = gr.Dropdown(
+                            label="src_lang",
+                            choices=[("English", "en"), ("Italian", "it")],
+                            show_label=False,
+                            filterable=True,
+                        )
+                        src_text = gr.Textbox(
+                            label="src_text",
+                            lines=8,
+                            interactive=True,
+                            show_label=False,
+                        )
+                        btn = gr.Button("Translate")
+
+                    with gr.Column():
+                        tgt_lang = gr.Dropdown(
+                            label="tgt_lang",
+                            choices=[("English", "en"), ("Italian", "it")],
+                            filterable=True,
+                            show_label=False,
+                        )
+                        tgt_text = gr.Textbox(
+                            label="tgt_text",
+                            placeholder="Translation will appear",
+                            lines=8,
+                            interactive=False,
+                            show_label=False,
+                            show_copy_button=True,
+                        )
+
+                    btn.click(
+                        _run_translation,
+                        inputs=[src_lang, src_text, tgt_lang],
+                        outputs=tgt_text,
+                    )
+
+        else:
+            raise ValueError(f"Type {type} not supported.")
+
         interface.launch()
